+++ module-apps/application-phonebook/data/PhonebookInternals.hpp
@@ -1,6 +1,22 @@
 #pragma once
 
+#include <cstdint>
+
 namespace phonebookInternals
 {
     constexpr uint8_t speedDialMaxValue = 9;
+
+    enum class ListItemName
+    {
+        FirstName,
+        SecondName,
+        Number,
+        OtherNumber,
+        Email,
+        SpeedDialKey,
+        AddToFavourites,
+        AddToICE,
+        Address,
+        Note
+    };
 } // namespace phonebookInternals